19419: Improves auto analyze threshold for TS cartpole (#23)
Seems that due to the extra features of time series, having it
auto-analyze too early causes the model to analyze on early/poorly
played games. Bumping up the auto_analyze threshold allows the
exploration aspect to train on more 'better' games prior to analyzing,
so that it can continue to improve at a faster rate and win faster than
before.